+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
 * BaseController: middleware方法改为public,允许外部(如通过事件)设置middleware
 * WeChatApp: 初步增加转发多客服功能
 * Record: 修复record有默认数据时,filter返回数据和默认数据混淆在一起的问题
+* Session: start方法支持session关闭后重新启动
 
 ## 0.9.20 (2015-06-18)